Removed legacy method.
@craigtaverner was fine with sorting out the problems via issues (if
they come up), so I removed the legacy method `isOneway(...)`. I did
think about making it private, but I feel like it would be a change
which is too difficult to sort out - the way it is now, a simple
renaming would be fine, while making the method private would require
ugly workarounds from the people who are actually using it by now. There
may be none, but I don't want to risk it.