Bulk load the RTree by @phil20686
With associated changes to the Layer interface to make it accessible through Layer.addNodes(List<Node>)

This is also used by the `spatial.addNodes('layer', nodes) yield count` procedure

Key Notes: created a much more efficient way to load an RTree, references for the algorithms given in javaDoc comments as appropriate.

Basic details below

1) if the insert large compared with the tree, rebuild the tree from scratch using a top down method which minimises overlapping.
2) Else use a seeded bulk clustering, i.e. use the existing tree structure to partition the data,
and then bulk insert by building an R tree from scratch and inserting its children at appropriate points.
3) Outliers and underpacked nodes are dealt with through one by one insertion at the end.
This means worst case behaviour is still polynomial if your data is constructed so that all inserts are always outliers.
4) This algorithm builds a tree with a loading factor of 0.7, this is slightly suboptimal from a querying point of view,
but enormously beneficial if you plan further inserts.

Incidentally, although this code works, there is a subtle flaw that makes the tree suboptimal.
Somewhere in the build tree function it splits up the incoming nodes.
iirc its in a function called partition which I added to the RTree file which attempts to presort the nodes,
but the method for partitioning doesnt respect the sorting. This means that the R tree that is created has very high overlap.
That was why this pull request was never merged into the main branch. I was going ot fix it but somehow I never got around to it.

Comment @jexp:
I changed that partition method to use list.subList() from left to right which should keep the ordering
Each sub-partition is sorted again then by x-min or y-min depending on depth
Hope that resolves the issue mentioned above