Cleanups including removal of unused code
Notably the get(Long) and get(Set<Long>) methods of LayerIndexReader and all implementations and usages of those methods. The Usage in Neo4jSpatialDataStore was itself broken and unusable anyway.

Also moved index specific interfaces and implementations into package org.neo4j.gis.spatial.index (ie. moved down one step to 'index'). This should also be considered for some other code as the spatial package is quite cluttered.

Also removed `@Procedure("spatial.updateFromWKT")` and the REST API version of this, because it was using the get(Long) interface above, which would be harder to implement on a non-in-graph index.