Major rework for porting to Neo4j 4.0
This commit was only to get it to compile against Neo4j 4.0, and involved changing almost every single file.
The next commit will be to get some or most tests to pass.
In addition:

* The largest work was supporting the new transactions model.
* Also removed support for the old REST API and SpatialPlugin.
* Added some stub code for a raw lucene alternative to the explicit indexes used for geohash and space-filling curves.
* Removed most uses of BatchInserter because of lack of lucene integration in Neo4j 4.0

The OSM importer was switched to online import only.
The plan is to integrate this with the external OSM library.
That will require unifying the two models, because the newer OSM library has evolved the model somewhat.

Work remaining:

* Suport lucene index for ExplicitIndexBackedPointIndex
* Port to use external OSM library for offline imports
* Fix all tests