Port to Neo4j 4.2.3
The change to Neo4j 4.2 was more subtle. Mostly only internal API's around the use of `Path` instead of `File`.
One change that could be noticed was the `IndexManager.IndexAccessMode` class.
In the 0.27.0 and 0.27.1 versions we used `OverridenAccessMode` to take the users existing access mode and simply add on the rights to create tokens and indexes. In 0.27.2 we instead use `RestrictedAccessMode` to restrict the users access right to the built in `AccessModel.Static.SCHEMA` and then boost to enable index and token writes. The difference is subtle and should only be possible to notice in Enterprise Edition.