Remove use of old reference nodes, reducing deadlocks
Also add support for exception on trying to import the same OSM file twice, as this is not yet supported and will cause errors or model corruptions.

This change is not backwards compatible, as the spatial model is different. Layers created in earlier verions would not be readible in this one. To protect against that we throw exceptions on all SpatialDatabaseService. layer finding methods, if they detect the old format. The exception requests that the database be upgraded.

This can be done in the Java API with the upgradeFromOldModel method, or via procedures using `spatial.upgrade`.