Use unique labels for indexing within multiple OSM layers
Essentially when adding nodes of a particular label, we also add a label with a unique name for indexing purposes.
The unique name is made of the original name plus a hex suffix made of the MD5 hash of the layer name.

For example, the layer `geom1` will have an MD5 hash of its name `9ECE5459EA0D46FC556E5E3F454A0795`.
Then when adding an OSM node we label the node with both:

* OSMNode
* OSMNode_9ECE5459EA0D46FC556E5E3F454A0795

The second label is also used to create a label-property index to be used for looking up the node when building ways.