New procedure spatial.merge.into(a,b) for merging one layer into another
This includes tests that pass for various combinations of simple point layers, but should also work for any layers that have property only geometry encoders.

There are failing tests for merging into OSM, which should not be supported, and for merging one OSM into another, which we hope to support in the next update.

In fact, this commit is preparatory work for OSM-OSM merging capabilities, which are planned, but not implemented yet.