Initial support for OSM layers in spatial.merge.into(a,b)
This required some refactoring of the OSM model with changes to OSMImporter, OSMLayer and OSMDataset.
In addition we added OSMModel for collecting general contants and functions appropriate to OSM, and we added OSMMerger with the intelligence to merge two OSM datasets together.

This is currently only tested in a trivial case, which is two identical models with one trivial change (added a free floating OSM node).

One change to the OSM model is that we moved User nodes to a linked list like we do for Ways and Relations.